pyproject.toml
[tool.poetry]
name = "svetka-py"
version = "0.1.0"
description = "Svetka AI - Desktop AI Assistant"
authors = ["Svetka Team"]
readme = "README.md"
packages = [{include = "svetka"}]

[tool.poetry.dependencies]
python = "^3.10"
PyQt6 = "^6.10.0"
PyQt6-WebEngine = "^6.10.0"
pydantic = "^2.12.5"
PyYAML = "^6.0.3"
sounddevice = "^0.5.3"
SoundCard = "^0.4.5"
opencv-python = "^4.12.0"
mss = "^10.1.0"
pygetwindow = "^0.0.9"
pyautogui = "^0.9.54"
pymsgbox = "^2.0.1"
pyscreeze = "^1.0.1"
pyrect = "^0.2.0"
pytweening = "^1.2.0"
keyboard = "^0.13.5"
psutil = "^7.2.1"
pyperclip = "^1.11.0"
edge-tts = "^7.2.7"
pynput = "^1.8.1"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
watchfiles = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.1"
pytest-qt = "^4.5.0"
pyinstaller = "^6.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=core --cov=modules --cov=ui --cov-report=term-missing"

[tool.coverage.run]
source = ["core", "modules", "ui"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
